Update Nerves dependencies/config to the latest
Despite the number of changes, all of these updates are fairly minor.
They bring in documentation updates, fix issues and update the
dependencies to the versions that would likely be used anyway. The only
visible change is the switch to NervesMOTD. The Nerves logo that gets
shown on boot will now have some general info printed underneath it.